Fixed export bar/added time format feature
There was concern that the export bar may not be hitting 100% whenever
the export button was clicked. I have changed the export featrure so it
will move to 100% so long as there is a response from the server. So
long as there is a connection, this should no longer be an issue.
Additionally I added a time format checkbox which allows the user to
change what format the information is displayed in; however, this may
not translate to the downloaded version of the final results.