Possilbe File Fix
Possibly fixed the file download issue. needs to be tested. Also updated check_duration function to test appropraitely for $("#race_time") so it does not throw an error after switching screens on account of the timeout of .5 seconds. If this fixes the file issue, the download_csv function can be deleted.